Geography

+++ PeliApi/DTO's/SalaDeCineCreacionDTO.cs
@@ -12,11 +12,11 @@ namespace PeliApi.DTO_s
         [Required]
         [StringLength(120)]
         public string Nombre { get; set; }
-        
-        
-        public float Latitud { get; set; }
-       
-        
-        public float Longitud { get; set; }
+
+        //[Range(-90, 90)]
+        public double Latitud { get; set; }
+
+       // [Range(-180, 180)]       
+        public double Longitud { get; set; }
     }
 }
